refactor: TSKSWCH macro
We don't need to register and code on the key pressed. we can just
flag that it was pressed and start the counter. If the counter as
elapsed to TAPPING_TERM we can register the key combination to open
the gnome overview and hold the ALT key. If the counter had not
reached the TAPPING_TERM before the key was depressed. We don't need
to unregister and code. this simplify the states.